performance improvements
With this speedup, when you have various pages with png images, the
improvements is significant.
As a example, I used an old dual core machine, with 104 png images, to
generate a pdf, 1 image per page. The time consuming before was 1 minute
and 28 seconds. After this patch, the time is 6 seconds!!!
Yes, 6 seconds, 14,7x faster!