Centering of images incorrectly offsets the x pos
When positioning an image and align is set to center a new x position
is calculated to display the image in the middle of the page.

The current behavior removes the left margin from the width of the page
before calculating the center for the given image. If the intent was to
take the margins in to account then it's ignoring the right margin.

In order to center an image in the page I believe the current value for
x should be ignored which is what this change does.

I have added a test that places a big image and makes it easy to see
what is happening but it can't be asserted as the write_html command
moves it to the next line and the position of the image can't be
found.